refactor: move export request notifications to a folder
