+++ apps/notifications/_config/notifications.config.ts
@@ -31,8 +31,8 @@ import { UnitKPIHandler } from '../_handlers/automatic/unit-kpi.handler';
 import { CollaboratorInviteHandler } from '../_handlers/innovations/collaborators/collaborator-invite.handler';
 import { CollaboratorUpdateHandler } from '../_handlers/innovations/collaborators/collaborator-update.handler';
 import { AccountDeletionHandler } from '../_handlers/innovations/delete-account/account-deletion.handler';
-import { ExportRequestFeedbackHandler } from '../_handlers/innovations/export-request-feedback.handler';
-import { ExportRequestSubmittedHandler } from '../_handlers/innovations/export-request-submitted.handler';
+import { ExportRequestFeedbackHandler } from '../_handlers/innovations/export-request/export-request-feedback.handler';
+import { ExportRequestSubmittedHandler } from '../_handlers/innovations/export-request/export-request-submitted.handler';
 import { InnovationStopSharingHandler } from '../_handlers/innovations/stop-sharing-innovation/innovation-stop-sharing.handler';
 import { InnovationWithdrawnHandler } from '../_handlers/innovations/withdraw-innovation/innovation-withdrawn.handler';
 import { MessageCreationHandler } from '../_handlers/messages/message-creation.handler';
